third_party/blink/web_tests/fast/css/getPropertyValue-column-rule-expected.txt

           

Bug 110698: getPropertyValue for -webkit-column-rule returns null, should compute the shorthand value

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS div1.style.getPropertyValue("-webkit-column-rule") is '10px solid black'
PASS div2.style.getPropertyValue("-webkit-column-rule") is '10px solid'
PASS div3.style.getPropertyValue("-webkit-column-rule") is '20px red'
PASS div4.style.getPropertyValue("-webkit-column-rule") is '10px'
PASS div5.style.getPropertyValue("-webkit-column-rule") is 'solid black'
PASS div6.style.getPropertyValue("-webkit-column-rule") is 'dashed'
PASS div7.style.getPropertyValue("-webkit-column-rule") is 'green'
PASS div8.style.getPropertyValue("-webkit-column-rule") is ""
NOTE: 'foo' is an illegal CSS value for '-webkit-column-rule'.
PASS div9.style.getPropertyValue("-webkit-column-rule") is '10px dotted black'
NOTE: If only few longhand properties are specified, getPropertyValue for shorthand property returns null.
PASS div10.style.getPropertyValue("-webkit-column-rule") is ""
PASS div11.style.getPropertyValue("-webkit-column-rule") is ""
PASS div12.style.getPropertyValue("-webkit-column-rule") is ""
PASS successfullyParsed is true

TEST COMPLETE
